feat: add viewer feature and camera text input group uses now mantine's text inputs.
